[STM,SPI,FIX] Correctly probe SPI Flash, when booting from SPI.
When booting directly from SPI Serial (NOR) Flashes, using the Serial
Flash Controller (SFC) on all current SoCs from STMicroelectronics, when
using the "FSM-Boot" mode, and also especially when storing U-Boot's
environments in SPI, the SPI device might not be correctly probed, when
subsequently initializing the SFC's FSM driver. This commit rectifies
this unwanted behaviour.

Fundamentally, empirical observation shows that it appears that the SFC
may not always transition correctly from "FSM-Boot"-mode (often, but
not always used during SPI boot) to the "FSM"-mode (typically used by
the U-Boot read/write "driver"). This is typically only a problem if the
system is actually booting directly from a SPI Serial (NOR) Flash, and
the "FSM-Boot"-mode has been enabled (with dual (x2) or quad (x4) mode)
by S/W (typically by the romgen-generated "poke-table"). In addition,
the use of the FSM driver is necessary to show this issue (i.e. this is
not a problem if the "Bit-Banging" driver is used instead).  Putatively,
the location and/or the size of U-Boot's environment variables may also
be a pertinent factor here.

It turns out this problem is due to a known H/W issue with the SFC
from STMicroelectronics - bug reference: RnDHV00053934. The issue is
that the FIFO is not guaranteed to be emptied when transitioning from
"FSM-Boot" mode to "FSM"-mode, and there may be some "stray" (1, 2 or 3)
bytes remaining in the FIFO. The recommend S/W work-around is to read
the FIFO until it is exhausted of whole 32-bit words. Then iteratively
insert into the FIFO a single 8-bit byte at a time, until such time as
there is a complete 32-bit word in the FIFO. At this juncture we can read
that single 32-bit word, logically emptying the FIFO, and sure in the
knowledge that there are effectively no "stray" bytes remaining the FIFO.

This commit was, in part, based on an original linux fix for the same
problem, written by Angus Clark <angus.clark@st.com>. Thanks Angus!

This was tested on a single B2020 board, and SPI was always correctly
probed, when booting both via GDB, and directly from SPI Serial Flash.
	B2020 (B2020C-02-R247), with STxH415 (Orly1, cut 2.x) fitted.

Testing was with Spansion S25FL128S part, and using ARM core (not ST40).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>