[STM,FIX] Allow the PBL to "chain-load" U-Boot with SMP on ARM
The recent commits to add support for linux SMP for ARM for SoCs from
STMicroelectronics, worked correctly if the system was booted *directly*
from FLASH.  However, SMP was not working correctly, if instead U-Boot
was downloaded+booted from GDB, or "chained" (as a secondary boot-loader)
from some other (primary) boot-loader (e.g. "PBL"). This commit is an
attempt to fix this specific issue.

Since release 2013.1 of the PBL, the PBL will now send all cores
(master+slaves) through the same primary entry point address, for
non-linux applications - such as U-Boot. Hence, when the latest version
of the PBL chains U-Boot, all the slave cores will start executing
U-Boot at the (correct virtual address for the) linker symbol "_start",
and not the FLASH (physical) address.

This commit modifies the behaviour of the low-level initialization to
treat all slave cores the exact same way - irrespective if U-Boot is
being booted directly from FLASH, or if it is being booted from SDRAM,
after having being placed there by some other boot-loader. That is, all
slaves will now be placed directly in the "First" (PIC) Holding-Pen, and
moved to the SRAM Holding-Pen, once the master has created it in SRAM.
As a result, SMP should now work when U-Boot is booted either directly
from FLASH, or if it is chained from the (latest version of the) PBL.

It should be noted, that this was not tested by directly using the
PBL. Instead, GDB was used to "emulate" the PBL's new boot behaviour by
forcing the PC on the slave core to be "_start", before starting both
cores. This was achieved by using the following "trick":

	Append the parameter ",smp_mode=1" to armv7-linux-gdb, and set
	the slave core's PC to be the same entry point as for the master
	core:

		(gdb) load
		(gdb) info thread
		(gdb) thread 2
		(gdb) set $pc = _start
		(gdb) info thread
		(gdb) continue

After booting linux from U-Boot, on the linux console, we now get:
…
[    0.180000] Brought up 2 CPUs
[    0.180000] SMP: Total of 2 processors activated (3981.31 BogoMIPS).
…

Note: U-Boot was wiped from the FLASH boot-device prior to this GDB test,
      to ensure there was no vestigial effects.

Note: First ran "make romgen" to generate the poke-table used.

This was tested on a single B2020 board, verifying that SMP was working
correctly, when booted both: directly from SPI serial FLASH; and via GDB
(emulating the PBL's behaviour) as above.

	B2020 (B2020C-02-R247), with STxH415 (Orly1, cut 2.x) fitted.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>