[STM] Confirm the PBL "chain-loads" U-Boot with SMP on ARM
The recent commit to add support for linux SMP for ARM for SoCs from
STMicroelectronics, when U-Boot was "chain-loaded" from a primary
boot-loader (such as the "PBL"), was not tested using the PBL, but was
tested using GDB instead.

This commit now confirms that such behaviour with the PBL directly
chain-loading U-Boot has now been successfully tested, albeit with
a caveat.

Since version 2013.1 of the PBL, the PBL was modified to send all
cores (master+slaves) thought the same primary entry point address,
for non-linux applications - such as U-Boot. This indeed was the case,
and is precisely how U-Boot expects to be invoked.

In addition to this "contract" between the PBL and U-Boot, there is
another "timing" constraint aspect we also need to consider.  If we are
"chained" from a primary boot-loader (e.g. the PBL), then the onus is
on that primary boot-loader to ensure that our timing assumptions are
satisfied. Failure to satisfy these timing constraints may result in
the system appearing to boot correctly, but subsequent boots of (say)
SMP-linux might not start any (or all) of the slave cores. The timing
assumptions are summarised as follows:

	"The slave cores should all start U-Boot *before*
	(or plesiosynchronously with) the master core."

Failure to comply with this critical constraint can introduce a
"race-condition" with the management of the Holding-Pens, resulting in
undefined behaviour. Caveat Emptor!

When this was tested in anger (with the PBL), it turns out that the
previous commits to U-Boot compiled with this contract, but a bug in
the PBL was disinterred, which meant the PBL was not complying with this
contract, and hence SMP booting of linux from U-Boot when chain-loaded
from the PBL failed!

This bug is present in the PBL from version 2013.1 through to version
2013.2 (inclusive). It is envisaged that it will be fixed in the next
release of the PBL, putatively called the "2013.2.1" release.

A temporary workaround for the bug in the PBL was applied, and re-tested,
and SMP booting was then found to be successful. Hence, this commit,
is just to record that this version of U-Boot is now known to be "good",
and it will work with a suitably fixed version of the PBL.

Thanks, in particular, to Benjamin Baratte <benjamin.baratte@st.com>,
for performing the actual testing, and helping to resolve this issue.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>