[STM] Add initial support for the STxH315-ADI (B2116) board
This commit adds initial support for the STxH315-ADI (B2116) board from
STMicroelectronics. This support has only had minimal testing, and only
using the ST40 core, and only tested via GDB download.

In addition, it appears that there is a board issue with the on-board
ethernet, in particular, I suspect an issue with pull-up for MDIO on GMAC!

Testing was limited to a single board:

	B2116-B02 (STiH315-ADIB5ZA-015)

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>