[STM] Enable internal pull-up on MDIO for B2116
This commit enables the internal pull-up resistor for the MDIO signal
connected to GMAC #1, on the B2116 (STxH315-ADI) board. That board
should have had a 1k5 pull-up resistor fitted (RE18) to the MDIO line,
but it was not fitted.

With this fix, MDIO now correctly yields 0xFFFF when reading from
non-addressable PHY addresses.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>