[STM] Add post-AN timeout for Realtek RTL8211E PHY
It appears that (empirically) the RealTek RTL8211E(G) PHY requires an
additional delay, after the auto-negotiation (AN) has completed, but
before the "Link Status" bitfield in the BMSR actually becomes asserted!

This commit adds an optional loop (with a time-out), to re-read the BMSR
until either the Link Status is asserted, or the time-out is exceeded.

To enable this additional loop, the following macro should be defined:
CONFIG_STMAC_LINK_STATUS_TIMEOUT. This macro should be defined with
the delay (in seconds) for the maximum required time-out. By default,
this macro is *not* defined. However, for the Realtek RTL8211E PHY,
it is defined (by default) as being 50 ms, i.e.:

	#define CONFIG_STMAC_LINK_STATUS_TIMEOUT CONFIG_SYS_HZ/20

Testing was limited to using Ethernet with a single board, with both a
100Mbit switch, and a 1Gigabit switch connected to it. The board was:

	B2116-B02 (STiH315-ADIB5ZA-015)

Networking now seems to be very reliable, with no issues noted.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>