[STM] Use correct TMU frequency on STxH416 (and STxH315) on ST40
On the STxH416 (Orly2) and the STxH315 (a variant of the STxH416)
SoCs, the TMU (as used on the ST40 core) is actually clocked at 200MHz
(previous SoCs used 100MHz).  As a result, the TMU clock was running
twice as fast as U-Boot actually thought it was.

This resulted in various timers/timeouts terminating prematurely,
including "sleep" taking half as long as it should. More concerning,
it meant things like PHY resets were only asserted for 50% of the time,
which could result in the PHY not being reset properly - which can cause
undefined behaviour with networking!

This commit will now use a frequency of 200MHz for the TMU clock when
CONFIG_STM_STXH416 is defined, but we still use a frequency of 100MHz
when CONFIG_STM_STXH415 is defined.

This was tested on various boards:
	B2000: with STxH415 and STxH416 SoCs
	B2020: with STxH415 and STxH416 SoCs
	B2116: with the STxH315 SoC

Testing comprised building for the ST40 core, and for each combination
running "sleep 30", and comparing the delay with the "wall-time". To a
first order of approximation, the observed delay was correct.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>