[STM] Add initial support for Flattened Device Trees for ST's ARM SoCs
This commit adds initial support for natively booting ARM linux kernels
which use the new "Flattened Device Trees" (FDT) paradigm, on SoCs from
STMicroelectronics, such as the STxH415 and the STxH416.

To enable support for booting with FDTs, the following configuration
macro must be defined: "CONFIG_OF_LIBFDT".  Enabling this configuration
option, also adds the "fdt" family of sub-commands to U-Boot, and also
disables the use of ATAGs, for passing boot parameters to the kernel.

Note: to boot with a FDT, ST U-Boot expects the environment variable
"fdtaddr" to be defined, and contain the address of a valid FDT binary
blob, when the "bootm" command is invoked. This may be set using either
"setenv", or preferably using the new "fdt addr <addr>" sub-command.

Using the "fdt" family of commands can also be used to "edit" the FDT
binary blob in situ, such as overwriting the default "bootargs". Note
that if the FDT binary blob needs to grow, then you can use the "fdt move"
sub-command for this purpose. See "help fdt" for more information.

To help users with this new paradigm, the following U-Boot commands may
be instructive (based on a board with 2GiB of SDRAM @ 0x40000000):

		U-Boot> nfs 48000000 /export/vmlinux.ub
		U-Boot> iminfo 48000000

		U-Boot> nfs 47000000 /export/vmlinux.dtb
		U-Boot> iminfo 47000000

		U-Boot> fdt addr 47000000
		U-Boot> fdt print /chosen bootargs
(optional)	U-Boot> fdt move 47000000 47800000 20000 ; setenv fdtaddr 47800000
		U-Boot> fdt set /chosen bootargs \"console=ttyAS0,115200 ...\"
		U-Boot> fdt print /chosen bootargs

		U-Boot> bootm 48000000

This has only be given minimal testing, on a single board:

	B2020-C00, with a STiH416 SoC (STiH416-ADICMZA-175)

Testing comprised using GDB to download U-Boot, and then downloading
an ARM kernel, and a FDT ("*.dtb") over NFS, updating the binary blob,
and booting linux natively on ARM. Finally, the "u-boot.bin" image was
burned into SPI FLASH, and booted directly from SPI, and successfully
booted native SMP linux with a FDT.

Note: Booting ARM+FDT from an ST40 core, was *not* tested, as additional
      code changes are still required for cross-booting with FDTs.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>