[FIX] After "fdt move", update the "fdtaddr" environment variable.
Following on from the previous commit, when the "fdt move" command was
use to move/grow a FDT binary blob, the internal address of the "active"
(or working) FDT was correctly updated, but *not* the address in the
corresponding environment variable "fdtaddr".

This commit fixes this omission, by ensuring that both are kept in sync,
when "fdt move" is used successfully. As a result, it is now no longer
necessary to run "setenv fdtaddr <addr>", after executing "fdt move".

This has only be given minimal testing, on a single board:
	B2020-C00, with a STiH416 SoC (STiH416-ADICMZA-175)

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>