[STM] Completed initial testing of B2116 with STxH315 (Orly2)
This commit signifies initial successful testing of U-Boot on a B2116B
board, using STMicroelectronics' STxH315 (Orly2) SoC, with both the ST40
and the ARM cores.

To be clear, this U-Boot can boot an SMP ARM linux kernel natively, with
a Flattened Device Tree (FDT) / Device Tree Blob ("*.dtb"). It should be
noted that there is no support for linux on the ST40 core on this board.

Specifically, the following were tested, and known to work:
	RS-232 Serial Console (J26 and J29)
	Ethernet: with 100Mbps (FDX) and  1000Mbps (FDX) hubs
	EHCI (USB 2.x), on 3 root ports
	NAND (BCH18)
	SPI Serial Flash (FSM, Bit-Banging)
	Boot-from-NAND (BCH18)
	Boot-from-SPI
	Boot an SMP ARM linux kernel (natively)
	I2C

Specifically, the following are untested, or known *not* to work:
	The "reset" command
	HoM (Hibernate-on-Memory)
	ST40 cross-booting a ARM kernel
	OHCI (USB 1.x)
	Hamming Controller (NAND) on ARM (it is okay on ST40).
	Ethernet with a 10Mbps hub.

All the above was tested on a single STxH315-ADI board:
	B2116-B02 (STiH315-ADIB5ZA-015)

It should be noted that to boot linux on this board, the "vmlinux.ub"
image used, was *not* generated using "make uImage" from the linux
build system - some build bug is assumed!  Instead, the ".ub" image was
constructed manually by using the "mkimage" tool directly.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>