[STM] Move STM-specific driver files into drivers/stm/*
In a similar vein to commit 6b6060b24ac1593efb87cb863aa54522528f883e,
this commit moves the legacy STMicroelectronics-specific driver files
that were not in a ST40-specific directory. These files were in a
driver class-specific directory, and therefore ARCH-agnostic. However,
in hindsight, it is probably better to put all the STM-specific driver
files into a single directory, primarily for maintenance purposes.

Specifically, the following files were moved:

	renamed:	drivers/net/*stm* -> drivers/stm/*
	renamed:	drivers/serial/*stm* -> drivers/stm/*
	renamed:	drivers/usb/host/*stm* -> drivers/stm/*

In addition "ehci-stm.c" was renamed to "stm-ehci-host.c" to make it
clearer that this is a "host-side" file (just in case!).

Tested only by compiling for some STMicroelectronics' boards.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>