[STM] Migrate ST SoCs to use new network API (CONFIG_NET_MULTI)
In preparation for re-basing the STMicroelectronics tree to a later
version of U-Boot, this commit migrates STMicroelectronics' SoCs for
the ethernet API, from the (about to be removed) "legacy" (or "single")
MAC API, to the newer (soon to be the only) "multiple" MAC API.

This commit simply defines CONFIG_NET_MULTI for all STMicroelectronics'
SoCs, and re-factors "stm-stmac.c" accordingly, exclusively to use the
newer network API. Support for the older API has effectively been removed
for the "stm-stmac.c" driver.

It should be noted that the function "board_eth_init()" should ideally
be defined for *each* board.  However, "cpu_eth_init()" can be used as
a generic fall-back, for all boards of a given CPU type.  In any event,
"board_eth_init()" will have a higher priority, so define it, if it is
required to override the SoC-specific function.

Despite the usage of the newer "multi" ethernet API, none of the current
STMicroelectronics boards can yet use more than one MAC concurrently.
This limitation is not a function of the updated "stm-stmac.c" driver,
but of the PIO muxing/routing configuration in the SoC-specific files
that have not yet been refactored to allow a plurally of MACs to be
enabled at the same time.  In particular the following functions would
need to be revised: cpu_eth_init(), and stx<xxxx>_configure_ethernet().

This was tested only on a single STxH315-ADI board:

	B2116-B02 (STiH315-ADIB5ZA-015)

Networking seemed fine, including boot-from-SPI, using the ARM core
on this board. As this is a significant commit, with minimal testing,
it is possible that some board/SoCs may no longer work due to a bug
introduced herein.  Only further testing and time will tell…

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>