[NFS] Stop premature "umount" when copying large files over NFS
Prior to this commit, if we got an old (timed-out) message, (i.e. one with
a smaller rpc_id) then we would still transition the NFS state-machine
to the "STATE_UMOUNT_REQ" state, and prematurely terminate the transfer,
which is clearly incorrect!

This commit tries to fix this when in the "STATE_READ_REQ" state, by
simply just "dropping" the old (expired) message, and remaining in the
same "STATE_READ_REQ" state, when we get a result code of NFS_RPC_DROP.

This problem was only really noticed when copying large files (e.g.
bigger than 20MiB) over NFS. It seems heavily dependant on "network
load". Running "ping -f -s1000" to the target board also seems to increase
the reproducibility (but not always).  However, it has been reported that
running a "logo" splash screen seems to make this problem much more likely
to occur - presumably by altering the effective bandwidth on the LMI bus
(sheer speculation)!

It is difficult with any certainty to reproduce the network traffic
that can cause this unlikely scenario to occur. However, at least one
"dropped" packet was caught, and the remainder of the transfer appeared
to complete successfully, with such a (similar) change present.  Clearly,
further testing is required to have any confidence on this solution!

This was tested only on a single STxH315-ADI board:

	B2116-B02 (STiH315-ADIB5ZA-015)

See also:

	https://bugzilla.stlinux.com/show_bug.cgi?id=31368
	ClearQuest: RnDHV00054103

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>