[STM] Enable internal PU on STxH205 for USB over-current PIO
For cost-reduction reasons, the external (on-board) pull-up resistor
for the USB over-current signal has been removed on some reference
design boards.  A S/W workaround for this is to enable the internal
(on SoC) pull-up resistor for each such USB over-current PIO signal.

This commit implements this fix on all boards with the STxH205/7 SoC,
by enabling the internal PU unconditionally for all boards with this SoC.

For further details, please see:

	https://bugzilla.stlinux.com/show_bug.cgi?id=35992
	https://bugzilla.stlinux.com/show_bug.cgi?id=35994

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>