[STM] Rename TEXT_BASE to CONFIG_SYS_TEXT_BASE
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

The main-line tree had renamed the (old) #define macro from "TEXT_BASE" to
(the newer) "CONFIG_SYS_TEXT_BASE". This commit recapitulates this change
for the STMicroelectronics' git tree, for the recent 'v2013.01' merge.

It should be noted that this tree still does *not* build (on either ST40
or ARM), and further work still needs to be done to get it to build.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>