[STM] Use "partial linking" to generate the libraries.
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

The main-line tree was changed to use (relocatable) object files ("*.o"),
instead of (static) archive files ("*.a") for libraries. In addition,
these libraries were previously created with the "ar" utility, whereas,
they are now created by calling the make macro "cmd_link_o_target", which
generates a relocatable object file, by "partially linking" the library.
This commit recapitulates this change for the STMicroelectronics' git
tree, for the recent 'v2013.01' merge.

It should be noted that this tree still does *not* build (on either ST40
or ARM), and further work still needs to be done to get it to build.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>