[STM] Use ".u_boot_list" as the section name (e.g. for commands).
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

The main-line tree was modified to use a different Elf section name for
the list of "commands" linked, from ".u_boot_cmd" to ".u_boot_list".
This commit recapitulates this change for the STMicroelectronics' git
tree, for the recent 'v2013.01' merge.

It should be noted that this tree still does *not* build (on either ST40
or ARM), and further work still needs to be done to get it to build.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>