[STM] Support new "multiple-serial" driver API for STMicroelectronics
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

The main-line tree was modified to allow a plurality of concurrent serial
drivers, and allowing each driver to "register" itself. As this is an
incompatible change of the API, then the previous STMicroelectronics'
serial drivers (i.e. "ASC", and "DTF") needed to be re-factored to comply
with this new paradigm. This commit does this, for these drivers.

We also removed the driver-specific functions "serial_puts()", and used
the generic "default_serial_puts()" function instead.

It should be noted that this tree still does *not* build (on either ST40
or ARM), and further work still needs to be done to get it to build.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>