[STM] Remove the field "bi_ip_addr" from the "bd_info" structure
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

The main-line tree was modified to remove the "bi_ip_addr" field from
the "bd_info" structure, and use the environment variable "ipaddr"
instead to define the IP address of the target board.  This commit
recapitulates this change for the STMicroelectronics' git tree, for the
recent 'v2013.01' merge.

It should be noted that this tree still does *not* build (on either ST40
or ARM), and further work still needs to be done to get it to build.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>