[STM] Fix compiler WARNINGS for STMicroelectronics' tree
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

This commit is a miscellany of disparate fixes that were (empirically)
required to get U-Boot to compile cleanly (i.e. fixes compiler warnings)
for both the ST40 and ARM SoCs from STMicroelectronics. These fixes are
presumably all required due to various changes in the main-line from the
'v2010.09' release to the 'v2013.01' release.

It should be noted that although with this commit, the tree now does build
(on both ST40 and ARM) *cleanly*, no attempt has been make to run U-Boot
on any target board.  Testing by compilation only (with no warnings).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>