[STM] Do not compile networking code, if no MAC/GMAC driver is present
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

The main-line tree was modified in such a way that it could fail to link
if either of the configuration macros CONFIG_CMD_NFS or CONFIG_CMD_PING
were defined, but there was no ethernet driver configured (at least in
the case of the STMicroelectronics' git tree).

Clearly, if no MAC/GMAC driver is to be included, then we should not
include *any* of the "networking" sub-modules, otherwise we may have a
fatal linking problem!  In this commit, we conveniently and consistently
un-define them all in the STMicroelectronics-specific configuration
file "include/stm/config.h". Any similar code in the board-specific
configuration files was also deleted, to avoid partial duplications.

The following macros are all now un-defined, if there is no MAC/GMAC
driver configured for SoCs from STMicroelectronics:

		CONFIG_CMD_NET
		CONFIG_CMD_NFS
		CONFIG_CMD_PING
		CONFIG_CMD_DHCP

It should be noted that this tree still does *not* build (on either ST40
or ARM), and further work still needs to be done to get it to build.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>