[Fix] set LC_ALL=C when running the make_u_boot_list rule
Ensure that when setting the environment variable LC_COLLATE to "C"
(to run the "sort" utility) that the environment variable LC_ALL does
not contradict it, as LC_ALL seems to dominate LC_COLLATE.

In the make definition of the "make_u_boot_list" rule (in helper.mk),
there is a descriptive comment which says:

>	It is imperative to use LC_COLLATE=C here ...

Unfortunately, it looks like when the "sort" utility is running, if the
user has exported the environment variable LC_ALL (as I had), then this
"imperative" can be violated, and U-Boot builds, but it is not built
correctly - e.g. no "commands" available!

This commit explicitly sets both LC_ALL and LC_COLLATE to "C", so there
is no contradiction, and the aforementioned imperative can be honoured.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>