[STM] Get environment variables on SPI working again.
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

This commit essentially re-factors the code to use the updated paradigm
for working with "environment variables", including the new deployment of
"hash-tables".  This is specifiably for SPI serial flash, which (please
note) still uses the old "EEPROM" interface on STMicroelectronics'
git tree.

In addition, this commit updates the STMicroelectronics-specific file
"drivers/stm/env_eeprom.c", to deal with boot-from-SPI properly again.
This needed some very subtle changes to the way env_get_char_spec()
was both called, and was implemented, to ensure it worked in all cases
(i.e. pre-relocation, and post-relocation of the environment variables).

Also added a cheeky optimization to speed up the call to env_import(),
which (surprisingly) makes many calls to getenv() (and friends),
whilst the environment has still not been relocated into the
hash-tables! However, a RAM version of the entire binary representation
of the environment variables is present, and can be used temporarily,
but only whilst the function env_relocate_spec() is being invoked.

This was not tested by booting directly from SPI, but it was tested on a
single (B2116B) board: using the ST40 core; with boot-from-SPI enabled;
downloading U-Boot via GDB; and the environment variables stored directly
in SPI serial flash.

Note: still need to ensure combinations involving NAND are not broken).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>