[STM] Important fixes to "nand dump"
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

This commit fixes two closely related fixes for "nand dump" sub-command,
in the STMicroelectronics' GIT tree.

The first one, is the re-instatement of the following single line:
		p = oobbuf;
in the function nand_dump(), which was accidentally "lost" in the
'v2013.01' merge (commit 7fa7e4418f1fccf678c9d2dd1b4245dd21a1c679),
in the STMicroelectronics' tree.  This lost line was originally added
in commit cfdae12f37ad126d1198f6c1ffa4ada9e69dd206.

The second one, is as follows. In U-Boot 'v2010.09', the NAND data
structures appended the OOB to the DATA buffer ("datbuf"). The function
stm_remap_bch_buffers() took advantage of this fact, and the fact that
nand_dump() also passed to stm_remap_bch_buffers() a "spare" OOB buffer
("oobbuf") which it could readily use.  However, with the aforementioned
recent re-base to 'v2013.01', this exploit could no longer be used, as
the OOB is no longer appended to the end of the data buffer.  The result
was the corruption of the internal malloc data structures - ouch!

A work-around was to set the environment variable "bch_raw".

This commit also re-factors the STMicroelectronics-specific function
stm_remap_bch_buffers() to comply to the latest conventions regarding
this, and uses an automatic buffer as a temporary buffer instead, whilst
re-mapping the DATA+ECC sections.

For further information about the reasons for this 2nd fix, please also
see commit cfdae12f37ad126d1198f6c1ffa4ada9e69dd206.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>