[STM] Fix printing frequencies in MHz (for "bdinfo")
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

This commit stops dividing the frequency by 1,000,000 before calling
the function print_mhz() to print it, as this function now expects its
argument to be in units of Hertz (Hz), and not Mega-Hertz (MHz).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>