[STM] Use board_early_init_f() for ASC initialization
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

As part of the improvements to ARM, the function board_init_f() can now be
executed directly in FLASH - i.e. *before* U-Boot has even been relocated
to RAM, or before the .bss is initialized, or global data is accessible!

This has the advantage that "printf()" can be called very early,
which clearly helps with debugging. However, it has the disadvantage
that that all the early initialization code (including bringing up the
serial console) has to run in a very restrictive environment, and must be
extremely careful as to what assumptions it can make - as the C run-time
environment initialization is still incomplete!

In order to support this new paradigm, for all STMicroelectronics'
boards we will now call board_early_init_f() from board_init_f(), which
will just configure the PIOs to route the serial console's UART (ASC)
to the appropriate PIO pins on the board. Essentially, we do the minimum
initialization we can, as early as possible, just to allow the serial
console to work.

It should be noted that the rest of the board-specific initialization
will nominally be performed by the function board_init(), which for
ARM systems, is called from board_init_r(), which is *after* U-Boot
has been relocated, and after the C run-time environment has been fully
initialized.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>