[STM] Move fields from "struct bd_info" to "struct global_data"
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

As part of the improvements to ARM, the function board_init_f() is now
called before the field "bd" (struct bd_info) in struct global_data is
initialized. Whereas, it is initialized prior to calling board_init_r().
This means that any code in (or called by) board_init_f() can *not*
now use the field "bd" in struct global_data.

Prior to this update, support specifically for STMicroelectronics'
SoCs had the following additional fields in "struct bd_info":

	bi_devid, bi_uart_frq, bi_tmu_frq, bi_ssc_frq.

These fields were all initialized very early. However, with the
initialization of the field "bd" now being deferred, this is unsafe,
and resulted in NULL pointer dereferences.

This commit resolves this issue, by moving these fields from "struct
bd_info", to be in "struct global_data", and also renaming them to:

	stm_devid, stm_uart_frq, stm_tmu_frq, stm_ssc_frq.

Note: The fields with "_tmu_" in them, are exclusive to the ST40.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>