[STM] Fix Boot-From-SPI for STMicroelectronics' SoCs
Following on from the recent (incomplete) "snapshot" of the merge of
the official 'v2013.01' release, into the STMicroelectronics' git tree,
this commit is one step (of many steps) towards the final completion of
that aforementioned snapshot commit.

It is now expected that all boot-from-FLASH ARM systems will call
board_init_f() whilst executing directly out of FLASH, before any
relocations, or even the C run-time environment is properly set up!

However, for some SoCs from STMicroelectronics, this is a problem,
as when booting specifically from SPI serial FLASH, the SPI Serial
Flash "boot-controller" on some of these SoCs do not support LD1 (8-bit
accesses) or LD2 (16-bit accesses) operations on the STbus. Hence, it is
not possible to call board_init_f() (and all of its callee functions),
if there are any "char" or "short" accesses to the SPI serial FLASH
"boot-controller".

To solve this problem, we will "temporary" relocate U-Boot from FLASH
to RAM, before we call board_init_f().  That is, we will perform
an additional "first copy" to RAM, if we are booting directly from
FLASH. For simplicity, and for consistency, we will use this technique
for all FLASH booting types (e.g. including boot-from-NAND), even if
this is not strictly necessary, for all SoCs from STMicroelectronics.

Note: After board_init_f() returns, we will still call the standard
function relocate_code(), which will copy the code again to RAM, but this
2nd copy will *also* perform the relocation "fix-ups".  Typically the
2nd copy will be to a different destination address, from the first copy.

This was tested only on a single STxH315-ADI board:

	B2116-B02 (STiH315-ADIB5ZA-015)

Testing comprised configuring the board to boot the ARM core from SPI,
and verifying that the board did boot from SPI, both via power-on,
or via pressing the "reset" button. No problems noted.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>