[STM] Rename prefix from "MPE41_*" to "STM_A9_*".
In preparation for adding new support for the STxH407 SoC from
STMicroelectronics, we rename the macro "MPE41_GLOBAL_TIMER_BASE"
to "STM_A9_GLOBAL_TIMER_BASE". This is because the prefix "MPE41"
is not applicable to the STxH407, hence a prefix denoting the "A9"
for STMicroelectronics is more appropriate.

Also we define a new macro "STM_A9_CONFIG_BASE" in the SoC-specific file,
as different SoCs may have different values for this base address.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>