[STM] Lower Vcore on the B2120, if it was increased for JTAG.
On the B2120 board from STMicroelectronics, to boot linux reliably
we may need to decrease the voltage of the A9 sub-system (i.e. Vcore)
to a nominal value of 1.10V.  This is due to an issue in the JTAG-A9
connection whereby it appears some samples need to be powered at 1.30V,
but this elevated voltage can create consequential problems for some IP
blocks (e.g. PL310 L2$ controller).

This is just a temporary software work-around for such samples.
Define the following macro if you have a sample affected by such an issue:

	#define CONFIG_MACH_STM_STXH407_A9SS_VCORE_HACK

For the time being, and by default this macro is *enabled*!

By enabling the above macro, then 2 changes will occur:

	1) When booted from FLASH, then Vcore will be set to 1.10V
	   (instead of 1.30V) whilst the ('romgen' generated) "pokes"
	   are being performed by the poke-interpreter.

	2) In the function checkboard(), Vcore will be lowered to 1.10V,
	   *if* it is currently greater than this voltage.

WARNING:  If "romgen" is re-run, then change #1 above will be lost,
unless care is taken to replace the poke to 0x09510004, by the source
code #if defined(CONFIG_MACH_STM_STXH407_A9SS_VCORE_HACK) ...  #endif

WARNING:  After the Vcore voltage is decreased to 1.10V, then JTAG
connections can be lost, and GDB will loose control of the target system.
If you want to use GDB/JTAG to debug U-Boot, then please #undef the above.

Testing was limited to a single board:

	B2120-A02 (STiH407-HDK-AM00-045)

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>