[STM] Add support for the N25Q512 SPI Serial Flash
This commit adds support for the Micron N25Q512 SPI Serial Flash device.

Unlike previous Micron devices, it is now mandatory to check the "Flag
Status Register" following all WRITE or ERASE operations, and if required,
also clear any "sticky" bits using the Clear Flag Status Register command.
This commit adheres to this new paradigm for the above device. It is
likely that other (newer) devices in the same family may also need to
be converted to this new paradigm - this should now be trivial.

This was tested using a single N25Q512A83GSF40X device, on a single
STMicroelectronics board (B2147-A00-E006), using an unreleased minimal
(ST40) "profile" for this board, full support for which will be added
later, once it is complete.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>