[STM] Enable use of RealTek RTL8367RB switch on B2120 board
This commit now completes the addition of the inaugural support of the
RealTek RTL8367RB ethernet switch on the B2120 board (utilizing the
STxH407 SoC) from STMicroelectronics.

On the STxH407 SoC there is only a single GMAC device (GMAC1) present.
On the B2120 board, it is connected to an on-board RealTek RTL8367RB
switch, which we use in RGMII mode, as a (dumb) "fixed PHY", connected
at giga-bit speeds.

This additional support has only had minimal testing, and only for the
following features:

	RS-232 Serial console (ASC)
	Gigabit Ethernet with the RealTek RTL8367RB switch
	Booting a ARM SMP Linux kernel (with a DTB).
	SPI Serial Flash (with FSM driver)
	SPI Serial Flash (with Bit-Banging driver)
	Boot from SPI Serial Flash
	I2C
	"make romgen"	(see recent commit about Vcore!)
	Commands: bdinfo, sleep, printenv, saveenv, ping, nfs.

In particular, there is no support (yet) for eMMC.

Testing was limited to a single board:

	B2120-A02 (STiH407-HDK-AM00-045)

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>